Move Models list into support package.
